simplify ucsdped implementation
